requirements.txt
pyrofork
TgCrypto
python-dotenv
motor